requirements.txt
apipkg==2.1.0
argcomplete==1.12.3
asn1crypto==1.4.0
atomicwrites==1.4.0
attrs==21.2.0
backports.zoneinfo==0.2.1
boto3==1.19.0
botocore==1.22.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
cfn-flip==1.3.0
chardet==4.0.0
charset-normalizer==2.0.7
click==7.1.2
click-plugins==1.1.1
cligj==0.7.2
coverage==6.0.2
# cryptography==3.4.7         # Version 35.0.0 breaks zappa=0.52.0 deployments. (yes, their versions went from 3.4.7, 3.4.8, then 35.0.0)
dateparser==1.1.0
DateTime==4.3
defusedxml==0.7.1
Deprecated==1.2.13
docutils==0.17.1
Dumper==1.2.0
durationpy==0.5
execnet==1.9.0
Fiona==1.8.20
Flask==2.0.2
Flask-Compress==1.10.1
Flask-Cors==3.0.10
flask-lambda-python36==0.1.0
flask-talisman==0.8.1
future==0.18.2
geojson==2.5.0
geomet==0.3.0
gitdb==4.0.7
gitdb2==4.0.2
GitPython==3.1.24
gunicorn==20.1.0
hjson==3.0.2
hypothesis==6.37.0
idna==3.3
importlib-metadata==4.8.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.1.0
kappa==0.6.0
kml2geojson==4.0.2
lambda-packages==0.20.0
libpagure==0.22
lxml==4.7.1
MarkupSafe==2.0.1
more-itertools==8.10.0
munch==2.5.0
packaging==21.0
pandas==1.3.4
pathlib2==2.3.6
pep517==0.12.0
pexpect==4.8.0
pip-tools==6.4.0
placebo==0.10.0
pluggy==1.0.0
ptyprocess==0.7.0
py==1.10.0
pycparser==2.20
PyGithub==1.55
PyJWT==2.3.0
pykml==0.2.0
# PyNaCl==1.4.0             # breaks zappa 0.52.0 (didn't check earlier versions, not sure if we need this)
# pyOpenSSL==21.0.0         # (Requires cryptography, which makes zappa throw)
pyparsing==2.4.7
pyproj==3.2.1
pyshp==2.1.3
pytest==6.2.5
pytest-automation==1.1.2
pytest-cov==3.0.0
pytest-forked==1.3.0
pytest-xdist==2.4.0
python-dateutil==2.8.2
python-gitlab==2.10.1
python-slugify==5.0.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2021.10.8
requests==2.26.0
requests-toolbelt==0.9.1
responses==0.18.0
s3transfer==0.5.0
scandir==1.10.0
scikit-learn==1.0.1                  # WARNING: 0.24.1 breaks ShorelineMask26 test
serverless-wsgi==3.0.0
Shapely==1.7.1
six==1.16.0
smmap==4.0.0
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
tzdata==2021.4
tzlocal==2.0.0                     # tzlocal.get_localzone() changed it's return type after this (No 'localize' attr)
urllib3==1.26.7
Werkzeug==2.0.2
WKTUtils==1.1.4
wrapt==1.13.2
zipp==3.6.0
zope.interface==4.7.2
numpy==1.21.3
